# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Smart Chess Coach API
  description: |
    Smart Chess Coach - AI-Powered Chess Learning Platform
    
    A comprehensive chess analysis and coaching platform that provides:
    - Game analysis with Stockfish engine
    - AI-powered personalized coaching
    - Player progression tracking
    - Mistake pattern analysis and learning
    - Voice-based coaching feedback
    - Journey dashboard with comprehensive stats
  version: 1.0.0
  contact:
    name: Smart Chess Coach Team
  license:
    name: Proprietary
servers:
  - url: https://api.smartchesscoach.com
    description: Production
  - url: http://localhost:8000
    description: Local development
...
paths:
  /api/auth/google/login:
    get:
      tags:
        - Authentication
      summary: Start Google OAuth login flow
      description: Redirects user to Google OAuth consent screen for authentication
      responses:
        '200':
          description: Google authentication URL provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  auth_url:
                    type: string
                    description: URL to redirect user for Google OAuth
        '500':
          description: Google OAuth not configured
  /api/auth/google/callback:
    get:
      tags:
        - Authentication
      summary: Handle Google OAuth callback
      description: Exchange authorization code for session token. Called by Google after user consent.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from Google
      responses:
        '302':
          description: Redirect to frontend dashboard with success
        '401':
          description: Invalid or failed authorization
        '500':
          description: Authentication failed
  /api/auth/session:
    post:
      tags:
        - Authentication
      summary: Create session from Emergent session ID (legacy)
      description: Exchange Emergent session_id for session_token. Used for backwards compatibility.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Missing session_id
        '401':
          description: Invalid session_id
  /api/auth/google/mobile:
    post:
      tags:
        - Authentication
      summary: Mobile Google authentication
      description: Authenticate mobile app users with Google access token from Google Sign-In SDK
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                access_token:
                  type: string
                  description: Google OAuth access token from mobile SDK
              required:
                - access_token
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  session_token:
                    type: string
        '400':
          description: Missing or invalid access token
        '401':
          description: Invalid Google access token
  /api/auth/demo-login:
    post:
      tags:
        - Authentication
      summary: Demo login for testing (localhost only)
      description: Create demo user account for testing without Google OAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
      responses:
        '200':
          description: Demo login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  session_token:
                    type: string
        '400':
          description: Invalid email format
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Retrieve authenticated user's profile information
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated or session expired
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate session and clear authentication cookies
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
          description: Unique user identifier
        email:
          type: string
          format: email
        name:
          type: string
        picture:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
        chess_com_username:
          type: string
          nullable: true
        lichess_username:
          type: string
          nullable: true
      required:
        - user_id
        - email
        - name
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token as Bearer token in Authorization header
security:
  - cookieAuth: []
  - bearerAuth: []